ai-dental-voice/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-multipart>=0.0.9

# Stage 02 RAG
faiss-cpu>=1.8.0
numpy>=1.26.0
tiktoken>=0.7.0
PyMuPDF>=1.24.7
python-docx>=1.0.1

